requirements.txt
pandas==2.1.4
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
lxml==4.9.3
cryptography==41.0.7
tabulate==0.9.0
prefect==2.10.0
great-expectations==0.18.0
dask==2023.10.0
streamlit==1.28.0
plotly==5.18.0
boto3==1.33.0
pytest==7.4.0
ruff==0.1.0
structlog==23.1.0